Renames fit->kundu_fit
