+++ tedana/metrics/__init__.py
@@ -1,7 +1,7 @@
 # emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil -*-
 # ex: set sts=4 ts=4 sw=4 et:
 
-from .fit import (
+from .kundu_fit import (
     dependence_metrics, kundu_metrics, get_coeffs, computefeats2
 )
